+++ spec/ParseInstallation.spec.js
@@ -7,7 +7,7 @@ const Config = require('../lib/Config');
 const Parse = require('parse/node').Parse;
 const rest = require('../lib/rest');
 const request = require('../lib/request');
-const { ErrorMessage } = require('./Errors/message');
+const { ErrorMessage } = require('../src/Errors/message');
 
 let config;
 let database;
